Remove unnecessary casts for print statement

+++ local_planner/src/nodes/star_planner.cpp
@@ -187,12 +187,12 @@ void StarPlanner::buildLookAheadTree() {
   path_node_origins_.push_back(0);
   tree_age_ = 0;
 
-  ROS_INFO("\033[0;35m[SP]Tree (%.0f nodes, %.0f path nodes, %.0f expanded) calculated in %2.2fms.\033[0m",
-           (double)tree_.size(), (double)path_node_positions_.size(), (double)closed_set_.size(),
-           (std::clock() - start_time) / (double)(CLOCKS_PER_SEC / 1000));
+  ROS_INFO("\033[0;35m[SP]Tree (%lu nodes, %lu path nodes, %lu expanded) calculated in %2.2fms.\033[0m", tree_.size(),
+           path_node_positions_.size(), closed_set_.size(),
+           static_cast<double>((std::clock() - start_time) / static_cast<double>(CLOCKS_PER_SEC / 1000)));
   for (int j = 0; j < path_node_positions_.size(); j++) {
-    ROS_DEBUG("\033[0;35m[SP] node %.0f : [ %f, %f, %f]\033[0m", (double)j, (double)path_node_positions_[j].x(),
-              (double)path_node_positions_[j].y(), (double)path_node_positions_[j].z());
+    ROS_DEBUG("\033[0;35m[SP] node %i : [ %f, %f, %f]\033[0m", j, path_node_positions_[j].x(),
+              path_node_positions_[j].y(), path_node_positions_[j].z());
   }
 }
 }
